+++ src/ts/AdMob/Views/AdMobView.ts
@@ -10,7 +10,7 @@ import {
 import { AdUnitContainer, Orientation } from 'Ads/AdUnits/Containers/AdUnitContainer';
 import { IGDPREventHandler } from 'Ads/EventHandlers/GDPREventHandler';
 import { ProgrammaticTrackingMetricName, ProgrammaticTrackingService } from 'Ads/Utilities/ProgrammaticTrackingService';
-import { AbstractPrivacy, IPrivacyHandler } from 'Ads/Views/AbstractPrivacy';
+import { AbstractPrivacy, IPrivacyHandlerView } from 'Ads/Views/AbstractPrivacy';
 import { Platform } from 'Core/Constants/Platform';
 import { ICoreApi } from 'Core/ICore';
 
@@ -37,7 +37,7 @@ export interface IAdMobEventHandler extends IGDPREventHandler {
 const AFMAClickStringMacro = '{{AFMA_CLICK_SIGNALS_PLACEHOLDER}}';
 const AFMADelayMacro = '{{AFMA_RDVT_PLACEHOLDER}}';
 
-export class AdMobView extends View<IAdMobEventHandler> implements IPrivacyHandler {
+export class AdMobView extends View<IAdMobEventHandler> implements IPrivacyHandlerView {
 
     private _campaign: AdMobCampaign;
     private _iframe: HTMLIFrameElement;
@@ -136,18 +136,10 @@ export class AdMobView extends View<IAdMobEventHandler> implements IPrivacyHandl
         }
     }
 
-    public onPrivacy(url: string): void {
-        // do nothing
-    }
-
     public onPrivacyClose(): void {
         if (this._privacy) {
             this._privacy.hide();
         }
-    }
-
-    public onGDPROptOut(optOutEnabled: boolean): void {
-        // do nothing
     }
 
     public onBackPressed() {