+++ src/ts/Ads/Utilities/ProgrammaticTrackingService.ts
@@ -5,7 +5,10 @@ import { DeviceInfo } from 'Core/Models/DeviceInfo';
 
 export enum ProgrammaticTrackingErrorName {
     TooLargeFile = 'too_large_file', // a file 20mb and over are considered too large
-    BannerRequestError = 'banner_request_error'
+    BannerRequestError = 'banner_request_error',
+
+    AuctionV4StartMissing = 'auction_v4_missing_start',
+    AuctionV5StartMissing = 'auction_v5_missing_start'
 }
 
 export enum ProgrammaticTrackingMetricName {
@@ -17,17 +20,7 @@ export enum ProgrammaticTrackingMetricName {
     BannerAdImpression = 'banner_ad_impression'
 }
 
-export interface IProgrammaticTrackingErrorData {
-    event: ProgrammaticTrackingErrorName;
-    platform: string;
-    osVersion: string;
-    sdkVersion: string;
-    adType: string;
-    seatId: number | undefined;
-}
-
-export interface IProgrammaticTrackingMetricData {
-    event: ProgrammaticTrackingMetricName | undefined;
+export interface IProgrammaticTrackingData {
     metrics: IProgrammaticTrackingMetric[] | undefined;
 }
 
@@ -66,8 +59,7 @@ export class ProgrammaticTrackingService {
         const platform: Platform = this._platform;
         const osVersion: string = this._deviceInfo.getOsVersion();
         const sdkVersion: string = this._clientInfo.getSdkVersionName();
-        const metricData: IProgrammaticTrackingMetricData = {
-            event: undefined,
+        const metricData: IProgrammaticTrackingData = {
             metrics: [
                 {
                     tags: [
@@ -91,8 +83,7 @@ export class ProgrammaticTrackingService {
     }
 
     public reportMetric(event: ProgrammaticTrackingMetricName): Promise<INativeResponse> {
-        const metricData: IProgrammaticTrackingMetricData = {
-            event: undefined,
+        const metricData: IProgrammaticTrackingData = {
             metrics: [
                 {
                     tags: [